Fix LocalId type in SpawnResult to match system int type
- Changed SpawnResult.LocalId from ushort to int to match EntityIdentity.LocalId
- Removed unsafe cast from int to ushort
- All tests still passing (347 total)
- CodeQL scan: 0 vulnerabilities

+++ server/Mmo.Server/EntityService/EntityService.cs
@@ -49,7 +49,7 @@ public class EntityService : IEntityService
             _log.Debug("Spawned entity {PersistentId} as LocalId {LocalId} in Zone {ZoneId}",
                 entity.PersistentId, localId, zoneId);
 
-            return Task.FromResult(new SpawnResult(true, (ushort)localId, zoneId));
+            return Task.FromResult(new SpawnResult(true, localId, zoneId));
         }
         catch (Exception ex)
         {
